refactor: modularize server and spec
