+++ src/generator/project/mod.rs
@@ -1,0 +1,5 @@
+mod generate;
+mod format;
+
+pub use generate::generate_project_from_spec;
+pub use format::format_project;
