+++ src/server/mod.rs
@@ -1,3 +1,5 @@
-mod server;
+mod request;
+mod response;
+mod service;
 
-pub use server::AppService;
+pub use service::AppService;
