+++ src/server/service.rs
@@ -1,0 +1,127 @@
+use super::request::{parse_request, ParsedRequest};
+use super::response::{write_handler_response, write_json_error};
+use crate::dispatcher::{Dispatcher, HandlerResponse};
+use crate::router::Router;
+use crate::security::{SecurityProvider, SecurityRequest};
+use crate::spec::SecurityScheme;
+use may_minihttp::{HttpService, Request, Response};
+use std::collections::HashMap;
+use std::io;
+use std::sync::{Arc, RwLock};
+
+#[derive(Clone)]
+pub struct AppService {
+    pub router: Arc<RwLock<Router>>,
+    pub dispatcher: Arc<RwLock<Dispatcher>>,
+    pub security_schemes: HashMap<String, SecurityScheme>,
+    pub security_providers: HashMap<String, Arc<dyn SecurityProvider>>,
+}
+
+impl AppService {
+    pub fn new(
+        router: Arc<RwLock<Router>>,
+        dispatcher: Arc<RwLock<Dispatcher>>,
+        security_schemes: HashMap<String, SecurityScheme>,
+    ) -> Self {
+        Self {
+            router,
+            dispatcher,
+            security_schemes,
+            security_providers: HashMap::new(),
+        }
+    }
+
+    pub fn register_security_provider(
+        &mut self,
+        name: &str,
+        provider: Arc<dyn SecurityProvider>,
+    ) {
+        self.security_providers.insert(name.to_string(), provider);
+    }
+}
+
+impl HttpService for AppService {
+    fn call(&mut self, req: Request, res: &mut Response) -> io::Result<()> {
+        let ParsedRequest { method, path, headers, cookies, query_params, body } =
+            parse_request(req);
+
+        let route_opt = {
+            let router = self.router.read().unwrap();
+            router.route(method.parse().unwrap(), &path)
+        };
+        if let Some(mut route_match) = route_opt {
+            route_match.query_params = query_params.clone();
+            if !route_match.route.security.is_empty() {
+                let sec_req = SecurityRequest {
+                    headers: &headers,
+                    query: &query_params,
+                    cookies: &cookies,
+                };
+                let mut authorized = false;
+                'outer: for requirement in &route_match.route.security {
+                    let mut ok = true;
+                    for (scheme_name, scopes) in &requirement.0 {
+                        let scheme = match self.security_schemes.get(scheme_name) {
+                            Some(s) => s,
+                            None => {
+                                ok = false;
+                                break;
+                            }
+                        };
+                        let provider = match self.security_providers.get(scheme_name) {
+                            Some(p) => p,
+                            None => {
+                                ok = false;
+                                break;
+                            }
+                        };
+                        if !provider.validate(scheme, scopes, &sec_req) {
+                            ok = false;
+                            break;
+                        }
+                    }
+                    if ok {
+                        authorized = true;
+                        break 'outer;
+                    }
+                }
+                if !authorized {
+                    write_json_error(
+                        res,
+                        401,
+                        serde_json::json!({"error": "Unauthorized"}),
+                    );
+                    return Ok(());
+                }
+            }
+            let is_sse = route_match.route.sse;
+            let handler_response = {
+                let dispatcher = self.dispatcher.read().unwrap();
+                dispatcher.dispatch(route_match, body, headers, cookies)
+            };
+            match handler_response {
+                Some(HandlerResponse { status, body }) => {
+                    write_handler_response(res, status, body, is_sse);
+                }
+                None => {
+                    write_json_error(
+                        res,
+                        500,
+                        serde_json::json!({
+                            "error": "Handler failed or not registered",
+                            "method": method,
+                            "path": path
+                        }),
+                    );
+                }
+            }
+        } else {
+            write_json_error(
+                res,
+                404,
+                serde_json::json!({"error": "Not Found", "method": method, "path": path}),
+            );
+        }
+        Ok(())
+    }
+}
