+++ src/spec/mod.rs
@@ -1,3 +1,8 @@
-mod spec;
+pub use oas3::spec::{SecurityRequirement, SecurityScheme};
+mod types;
+mod build;
+mod load;
 
-pub use spec::*;
+pub use build::*;
+pub use load::*;
+pub use types::*;
